% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareGeneset.R
\name{prepareGeneset}
\alias{prepareGeneset}
\title{prepareGeneset}
\usage{
prepareGeneset(db.type, Tissue, Species = "Human", Cancer = TRUE)
}
\arguments{
\item{db.type}{scType, CellMarker, CellTaxonomy, or snRandom-seq}

\item{Species}{Human or Mouse}

\item{Cancer}{TRUE or FALSE}

\item{db}{reference cell marker}
}
\description{
prepareGeneset
}
