% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAnnotation.R
\name{scAnnotation}
\alias{scAnnotation}
\title{scAnnotation}
\usage{
scAnnotation(
  seurat_obj,
  annot.method = "marker",
  dataset_ref = NULL,
  db.type = "CellMarker",
  Tissue,
  Species = "Human",
  Cancer = TRUE
)
}
\arguments{
\item{annot.method}{dataset or marker}

\item{dataset_ref}{If annot.method = 'dataset', you should
provide a SingleR reference dataset in `.rds` format.}

\item{db.type}{scType, CellMarker, CellTaxonomy, snRandom.
Only CellMarker and CellTaxonomy support Mouse.}

\item{Cancer}{}
}
\description{
scAnnotation
}
